backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
python-multipart==0.0.20
pandas==2.3.3
numpy==2.0.2
openpyxl==3.1.5
xlrd==2.0.2
python-dotenv==1.0.0
pydantic==2.9.2

# AI providers
groq>=0.4.0
google-generativeai>=0.3.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality tools
black==24.1.1
ruff==0.1.9
pre-commit==3.6.0

# Rate limiting
slowapi==0.1.9

# PDF generation
reportlab==4.0.7
Pillow==10.2.0

# Encryption for shareable links
cryptography==41.0.7
